Appropriately handle responses with no content.
When there is no response body present ResultSet is created using an empty hash.

As I understand the JSON API spec (see http://jsonapi.org/format/#crud-deleting as an example) returning no content in the body is a perfectly valid thing to do in a number of instances. Previously this would cause errors when trying to construct a ResultSet instance from the response.